位置情報取得の機能を useCurrentPosition.ts に切り出し
