+++ src/components/SunsetInfo/CurrentLocation.tsx
@@ -1,39 +1,10 @@
-import { useEffect, useState, VFC } from 'react';
+import { VFC } from 'react';
 import { Box, HStack, Text } from '@chakra-ui/react';
 
+import { useCurrentPosition } from '../../hooks/useCurrentPosition';
+
 export const CurrentLocation: VFC = () => {
-  const [lat, setLat] = useState(0); // Latitude 緯度
-  const [lon, setLon] = useState(0); // Longitude 経度
-  const [locationError, setLocationError] = useState('');
-
-  const getCurrentPosition = () => {
-    return new Promise(
-      (
-        resolve: (value: GeolocationPosition) => void,
-        reject: (reason: GeolocationPositionError) => void,
-      ) => {
-        navigator.geolocation.getCurrentPosition(resolve, reject);
-      },
-    );
-  };
-
-  useEffect(() => {
-    if ('geolocation' in navigator) {
-      const getPosition = async () => {
-        try {
-          const position = await getCurrentPosition();
-          setLat(position.coords.latitude);
-          setLon(position.coords.longitude);
-        } catch (err) {
-          if (err instanceof GeolocationPositionError) {
-            setLocationError(err.message);
-          }
-        }
-      };
-
-      getPosition();
-    }
-  }, []);
+  const { lat, lon, locationError } = useCurrentPosition();
 
   return (
     <Box
